% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHist3D.R
\name{buildHist3D}
\alias{buildHist3D}
\title{Title}
\usage{
buildHist3D(
  data,
  nbins = 15,
  bin.width = 0.4,
  xAxis.label = "x",
  yAxis.label = "y",
  zAxis.label = "",
  smooth.factor = 3,
  aspect.ratio = list(x = 1, y = 1, z = 1.5),
  xAxis.legend = TRUE,
  yAxis.legend = TRUE,
  zAxis.legend = FALSE,
  xAxis.tickangle = 0,
  yAxis.tickangle = 0,
  zAxis.tickangle = 0,
  axis.fontsize = "14px",
  legend.font = "Arial",
  legend.valign = "top",
  color.palette = "Viridis",
  caption = NULL,
  plot.title = NULL,
  title.fontsize = "24px",
  title.font = "Arial"
)
}
\arguments{
\item{data}{data.table}

\item{nbins}{integer}

\item{bin.width}{numeric}

\item{xAxis.label}{string}

\item{yAxis.label}{string}

\item{zAxis.label}{string}

\item{smooth.factor}{integer}

\item{aspect.ratio}{list}

\item{xAxis.legend}{boolean}

\item{yAxis.legend}{boolean}

\item{zAxis.legend}{boolean}

\item{xAxis.tickangle}{numeric}

\item{yAxis.tickangle}{numeric}

\item{zAxis.tickangle}{numeric}

\item{axis.fontsize}{string}

\item{legend.font}{string}

\item{legend.valign}{string}

\item{color.palette}{palette}

\item{caption}{string}

\item{plot.title}{string}

\item{title.fontsize}{string}

\item{title.font}{string}
}
\value{
plotly object
}
\description{
Title
}
