% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlot.ggplot2.R
\name{buildPlot.ggplot2}
\alias{buildPlot.ggplot2}
\title{plot.ggplot2}
\usage{
buildPlot.ggplot2(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{extra arguments}
}
\value{
ggplot2 object
}
\description{
plot.ggplot2
}
