% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlot.highcharter.R
\name{buildPlot.highcharter}
\alias{buildPlot.highcharter}
\title{Title}
\usage{
buildPlot.highcharter(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{extra arguments}
}
\value{
highcharter object
}
\description{
Title
}
