% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlot.Model.R
\name{buildPlot.Model}
\alias{buildPlot.Model}
\title{Build a model plot}
\usage{
buildPlot.Model(
  data.lines,
  data.points,
  xAxis.legend,
  yAxis.legend,
  line.width = 1.5,
  point.size = 2,
  point.shape = "circle"
)
}
\arguments{
\item{data.lines}{A data.table with at least three columns: X (numeric), Y (numeric), and ID (factor).}

\item{data.points}{A data.table with at least three columns: X (numeric), Y (numeric), and ID (factor).}

\item{xAxis.legend}{The title for the x-axis.}

\item{yAxis.legend}{The title for the y-axis.}

\item{line.width}{The width of the line.}

\item{point.size}{The size of the points.}

\item{point.shape}{The shape of the points.}
}
\value{
A highchart object representing the model plot.
}
\description{
This function creates a model plot with data lines and points.
}
