% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2D.R
\name{hist2D}
\alias{hist2D}
\title{Title}
\usage{
hist2D(
  data,
  nbins = 30,
  xAxis.label = "x",
  yAxis.label = "y",
  zAxis.label = "z",
  xAxis.min = NULL,
  xAxis.max = NULL,
  yAxis.min = NULL,
  yAxis.max = NULL,
  axis.fontsize = "14px",
  legend.font = "Arial",
  color.palette = hcl.pals()[33],
  plot.type = c("heatmap", "contour"),
  plot.title = NULL,
  title.fontsize = "24px",
  title.font = "Arial"
)
}
\arguments{
\item{data}{data.table}

\item{nbins}{integer}

\item{xAxis.label}{string}

\item{yAxis.label}{string}

\item{zAxis.label}{string}

\item{xAxis.min}{numeric}

\item{xAxis.max}{numeric}

\item{yAxis.min}{numeric}

\item{yAxis.max}{numeric}

\item{axis.fontsize}{string}

\item{legend.font}{string}

\item{color.palette}{palette}

\item{plot.type}{string}

\item{plot.title}{string}

\item{title.fontsize}{string}

\item{title.font}{string}
}
\value{
plotly object
}
\description{
Title
}
