% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showASCII.R
\name{showASCII}
\alias{showASCII}
\title{Display text file as plain ASCII (non-formatted)}
\usage{
showASCII(filePath)
}
\arguments{
\item{filePath}{Path to the text file}
}
\value{
Prints content using HTML pre tag for results: asis
}
\description{
Display text file as plain ASCII (non-formatted)
}
