% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMarkdownRendered.R
\name{showMarkdownRendered}
\alias{showMarkdownRendered}
\title{Display markdown file as rendered HTML via iframe}
\usage{
showMarkdownRendered(
  filePath,
  height = "500px",
  theme = "spacelab",
  toc = FALSE
)
}
\arguments{
\item{filePath}{Path to markdown file (will be rendered to HTML at runtime)}

\item{height}{Height of iframe (default "500px")}

\item{theme}{HTML theme (default "spacelab")}

\item{toc}{Show table of contents (default FALSE)}
}
\value{
Prints iframe HTML using cat() for results: asis
}
\description{
Display markdown file as rendered HTML via iframe
}
