% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlot.Histogram.R
\name{buildPlot.Histogram}
\alias{buildPlot.Histogram}
\title{Build a histogram plot}
\usage{
buildPlot.Histogram(
  .data,
  xTitle = "",
  title = "",
  histogram.breaks = 100,
  logscale = FALSE,
  offset = 0.1
)
}
\arguments{
\item{.data}{A data.table with at least two columns: X (numeric) and ID (factor).}

\item{xTitle}{The title for the x-axis.}

\item{title}{The main title for the plot.}

\item{histogram.breaks}{The number of breaks for the histogram.}

\item{logscale}{Logical, whether to use a logarithmic scale for the x-axis.}

\item{offset}{The percentage of the range to extend the x-axis.}
}
\value{
A highchart object representing the histogram plot.
}
\description{
This function creates a histogram plot with density curves and statistical markers.
}
