% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.R
\name{buildTable}
\alias{buildTable}
\title{Render a table with specified formatting options}
\usage{
buildTable(
  .x,
  library = "gt",
  format = "html",
  font.size.header = 14,
  font.size.body = 12,
  font.family.header = "Arial",
  font.family.body = "Arial",
  caption = NULL,
  font.bold.header = TRUE,
  font.bold.body = FALSE,
  font.bold.all = NULL,
  font.size.all = NULL,
  font.family.all = NULL,
  vlines.show = FALSE,
  hlines.show = TRUE,
  vlines.color = "grey",
  hlines.color = "grey",
  vlines.size = 1,
  hlines.size = 1,
  align.header = "center",
  align.body = "left"
)
}
\arguments{
\item{.x}{A data frame or data table to be rendered as a table.}

\item{library}{The library to be used for rendering. Options are "flextable", "gt", and "kable".}

\item{format}{The output format. Options are "html", "pdf", and "docx".}

\item{font.size.header}{The font size for the table header.}

\item{font.size.body}{The font size for the table body.}

\item{font.family.header}{The font family for the table header.}

\item{font.family.body}{The font family for the table body.}

\item{caption}{The table caption.}

\item{font.bold.header}{Logical. Whether to bold the header font.}

\item{font.bold.body}{Logical. Whether to bold the body font.}

\item{font.bold.all}{Logical. Whether to bold all text. Overrides font.bold.header and font.bold.body.}

\item{font.size.all}{Numeric. Font size for all text. Overrides font.size.header and font.size.body.}

\item{font.family.all}{Character. Font family for all text. Overrides font.family.header and font.family.body.}

\item{vlines.show}{Logical. Whether to show vertical lines.}

\item{hlines.show}{Logical. Whether to show horizontal lines.}

\item{vlines.color}{Color of vertical lines.}

\item{hlines.color}{Color of horizontal lines.}

\item{vlines.size}{Thickness of vertical lines.}

\item{hlines.size}{Thickness of horizontal lines.}

\item{align.header}{Alignment of header text. Options are "center", "left", "right".}

\item{align.body}{Alignment of body text. Options are "center", "left", "right".}
}
\value{
A formatted table rendered using the specified library and options.
}
\description{
This function renders a table using the specified library and formatting options.
It supports three libraries: flextable, gt, and kable.
}
\examples{

iris |> buildTable(library="gt")
iris |> buildTable(library="flextable")
iris |> buildTable(library="kable")

}
